requirements.txt
langchain
langchain-community
ollama
langchain_chroma
langchain_ollama
langchain-huggingface
wikipedia
openai
langchain_openai
Wikipedia-API
python-dotenv

